Depth layer converted to Single
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@11790 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/EwECore/Database/Updates/Release 6.3/cDBUpdate6_02_00_06.vb
@@ -65,7 +65,7 @@ Friend Class cDBUpdate6_02_00_06
         Dim info As cRegionInfo = Nothing
         Dim readerScenario As IDataReader = db.GetReader("SELECT ScenarioID, InRow, InCol, DepthMap FROM EcospaceScenario")
         Dim readerRegions As IDataReader = Nothing
-        Dim aiDepth(,) As Integer = Nothing
+        Dim aiDepth(,) As Single = Nothing
         Dim aiRegions(,) As Integer = Nothing
         Dim aiRegion(,) As Integer = Nothing
         Dim InRow, InCol As Integer
